+++ R/get_power.R
@@ -9,9 +9,9 @@
 #' @examples
 #' study() %>%
 #'  add_hypothesis("H1") %>%
-#'  add_analysis("A1", t.test(y~A, data = D1)) %>%
+#'  add_analysis("A1", t.test(y~B1, data = D1)) %>%
 #'  add_criterion("C1", "p.value", "<", 0.05) %>%
-#'  add_analysis("A2", t.test(y~A, data = D2)) %>%
+#'  add_analysis("A2", t.test(y~B1, data = D2)) %>%
 #'  add_criterion("C2", "p.value", "<", 0.05) %>%
 #'  add_eval("corroboration", "C1 & C2") %>%
 #'  add_eval("falsification", "!C1 & !C2") %>%