+++ R/study_from_xml.R
@@ -7,6 +7,12 @@
 #'
 #' @return A study object with class scivrs_study
 #' @export
+#'
+#' @examples
+#' grobid_dir <- system.file("grobid", package="scienceverse")
+#' filename <- file.path(grobid_dir, "eyecolor.pdf.tei.xml")
+#' study <- study_from_xml(filename)
+#'
 study_from_xml <- function(filename, xml_type = c("auto", "grobid")) {
   xml_type <- match.arg(xml_type)
 
@@ -23,6 +29,8 @@ study_from_xml <- function(filename, xml_type = c("auto", "grobid")) {
     return(s)
   }
 
+  message("Processing ", basename(filename), "...")
+
   if (!file.exists(filename)) {
     stop("The file ", filename, " does not exist.")
   }
@@ -30,6 +38,8 @@ study_from_xml <- function(filename, xml_type = c("auto", "grobid")) {
   xml <- tryCatch(xml2::read_xml(filename), error = function(e) {
     stop("The file ", filename, " could not be read as XML")
   })
+
+
 
   # deal with XML types----
   # TODO: support more than grobid?
@@ -48,11 +58,11 @@ study_from_xml <- function(filename, xml_type = c("auto", "grobid")) {
     xlist <- xml2::as_list(xml)
     s <- study()
 
+    s$name <- basename(filename)
     s$info$title <- xlist$TEI$teiHeader$fileDesc$titleStmt$title[[1]]
-    s$name <- s$info$title
 
     # abstract ----
-    s$info$abstract <- xlist$TEI$teiHeader$profileDesc$abstract |>
+    s$info$description <- xlist$TEI$teiHeader$profileDesc$abstract |>
       unlist() |>
       paste(collapse = " ") |>
       trimws()
@@ -66,7 +76,7 @@ study_from_xml <- function(filename, xml_type = c("auto", "grobid")) {
       given <- a$persName$forename[[1]]
       email <- a$email[[1]]
       orcid <- a$idno[[1]]
-      if (is.null(orcid)) {
+      if (is.null(orcid) & !is.null(family)) {
         orcid_lookup <- get_orcid(family, given)
         if (length(orcid_lookup) == 1) orcid <- orcid_lookup
       }
@@ -75,9 +85,19 @@ study_from_xml <- function(filename, xml_type = c("auto", "grobid")) {
     }
 
     # process text----
+    abstract <- xlist$TEI$teiHeader$profileDesc$abstract
+    abst_table <- full_text_table_from_grobid(abstract)
+    abst_table$section_class <- "abstract"
+    abst_table$section <- "div_0"
+    abst_table$div <- 0
+    abst_table$tag <- gsub("div_1", "div_0", abst_table$tag)
+
     body <- xlist$TEI$text$body
-    ft <- full_text_table_from_grobid(body)
-    s$full_text <- ft
+    body_table <- full_text_table_from_grobid(body)
+
+    s$full_text <- rbind(abst_table, body_table)
+    s$full_text$section_class <- factor(s$full_text$section_class,
+                                        levels = unique(s$full_text$section_class))
   } else {
     stop("This function cannot yet handle an XML of type ", xml_type)
   }
@@ -144,16 +164,16 @@ full_text_table_from_grobid <- function(body) {
 
   # make table of sections and classify
   headers <- ft[grepl("head", ft$tag), c("section", "text")]
-  missing_sections <- data.frame(
-    section = unique(ft$section) %>% setdiff(headers$section),
-    text = ""
-  )
-  sections <- rbind(headers, missing_sections)
+  sections <- data.frame(
+    section = unique(ft$section)
+  ) %>%
+    merge(headers, by = "section", all.x = TRUE)
+
 
   intro <- grepl("intro", sections$text, ignore.case = TRUE)
   method <- grepl("method", sections$text, ignore.case = TRUE)
   results <- grepl("result", sections$text, ignore.case = TRUE)
-  discussion <- grepl("discuss", sections$textaders, ignore.case = TRUE)
+  discussion <- grepl("discuss", sections$text, ignore.case = TRUE)
   sections$section_class[intro] <- "intro"
   sections$section_class[method] <- "method"
   sections$section_class[results] <- "results"
@@ -161,16 +181,96 @@ full_text_table_from_grobid <- function(body) {
 
   # assume sections are the same class as previous if unclassified
   for (i in seq_along(sections$section_class)) {
-    if (is.na(sections$section_class[i]))
+    if (is_nowt(sections$section_class[i]) & i > 1)
       sections$section_class[i] <- sections$section_class[i-1]
+  }
+
+  # replace blank sections with header text
+  blanks <- is_nowt(sections$section_class)
+  sections$section_class[blanks] <- sections$text[blanks]
+
+  # beginning sections with no header labelled intro
+  non_blanks <- which(!is_nowt(sections$section_class))
+  if (length(non_blanks) > 0) {
+    blank_start <- non_blanks[[1]] - 1
+    blanks <- rep(c(TRUE, FALSE), c(blank_start, length(blanks) - blank_start))
+    sections$section_class[blanks] <- "intro"
   }
 
   # add last to override non-div sections
   sec_labels <- gsub("_\\d+$", "", sections$section)
-  sections$section_class[sec_labels != "div"] <- sec_labels[sec_labels != "div"]
+  notdivs <- sec_labels != "div"
+  sections$section_class[notdivs] <- sec_labels[notdivs]
+  tables <- grepl("table", sections$text, ignore.case = TRUE)
+  sections$section_class[notdivs & tables] <- "table"
 
   # add sections to full text
-  ft <- dplyr::left_join(ft, sections[, c("section", "section_class")], by = c("section"))
+  ft <- merge(ft, sections[, c("section", "section_class")],
+              by = "section", all.x = TRUE)
+
+  # TODO: process table sections
 
   return(ft)
+}
+
+#' Search the full text
+#'
+#' @param study a study object created by `study_from_grobid` or a list of study objects
+#' @param term the regex term to search for
+#' @param section the section(s) to search in
+#' @param refs whether to include references
+#' @param ignore.case whether to ignore case when text searching
+#' @param ... additional arguments to pass to `grepl()`
+#'
+#' @return a data frame of matching sentences
+#' @export
+#'
+#' @examples
+#' grobid_dir <- system.file("grobid", package="scienceverse")
+#' filename <- file.path(grobid_dir, "eyecolor.pdf.tei.xml")
+#' study <- study_from_xml(filename)
+#' sig <- search_full_text(study, "significant", "results")
+search_full_text <- function(study, term, section = NULL, refs = FALSE, ignore.case = TRUE, ...) {
+  section_class <- text <- div <- p <- s <- NULL
+
+  # handle list of scivrs objects ----
+  if (!"scivrs_study" %in% class(study)) {
+    contains_scivrs <- lapply(study, class) |>
+      sapply(\(x) "scivrs_study" %in% x)
+    if (all(contains_scivrs)) {
+      matches <- lapply(study, \(x) {
+        search_full_text(x, term, section, refs, ignore.case, ...)
+      })
+      matches_agg <- do.call(rbind, matches)
+      return(matches_agg)
+    } else {
+      stop("The study argument doesn't seem to be a scivrs_study object or a list of study objects")
+    }
+  }
+
+  # filter full text
+  section_filter <- seq_along(study$full_text$section_class)
+  if (!is.null(section))
+    section_filter <- study$full_text$section_class %in% section
+  ref_filter <- TRUE
+  if (!refs) ref_filter <- study$full_text$type != "ref"
+  ft <- study$full_text[section_filter & ref_filter, ]
+
+  # get all sentences with at least 1 part matching term
+  match_term <- grepl(term, ft$text, ignore.case = ignore.case, ...)
+  ft_match <- ft[match_term, ]
+  # adds back the other sentence parts for divided sentences
+  ft_match_all <- dplyr::semi_join(ft, ft_match, by = c("div", "p", "s"))
+
+  full_text_table <- dplyr::group_by(ft_match_all, section_class, section, div, p, s) %>%
+    dplyr::summarise(text = paste(text, collapse = ""))
+
+  # full_text_table <- aggregate(text ~ section_class + section + div + p + s,
+  #                              data = ft_match_all,
+  #                              FUN = paste, collapse = "")
+
+
+  full_text_table$study <- study$name
+
+  return(full_text_table)
 }